redesign onboarding and login
